+++ internal/app/config_test.go
@@ -3,23 +3,38 @@ package app_test
 import (
 	"testing"
 
+	"github.com/pkg/errors"
+
 	"github.com/ReasonSoftware/security-group-manager/internal/app"
+	"github.com/ReasonSoftware/security-group-manager/mocks"
 	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/service/secretsmanager"
 	"github.com/stretchr/testify/assert"
 )
+
+const secret string = "secret"
 
 func TestGetConfig(t *testing.T) {
 	assert := assert.New(t)
 
 	type test struct {
-		Config         string
+		Parameter      string
+		MockInput      *secretsmanager.GetSecretValueInput
+		MockOutput     *secretsmanager.GetSecretValueOutput
+		MockError      error
 		ExpectedError  string
 		ExpectedOutput *app.Config
 	}
 
 	suite := map[string]test{
 		"Success": {
-			Config: `{
+			Parameter: secret,
+			MockInput: &secretsmanager.GetSecretValueInput{
+				SecretId:     aws.String(secret),
+				VersionStage: aws.String("AWSCURRENT"),
+			},
+			MockOutput: &secretsmanager.GetSecretValueOutput{
+				SecretString: aws.String(`{
 	"protocols": {
 		"http": {
 			"transport": "tcp",
@@ -42,7 +57,9 @@ func TestGetConfig(t *testing.T) {
 			"note": "London Office"
 		}
 	]
-}`,
+}`),
+			},
+			MockError:     nil,
 			ExpectedError: "",
 			ExpectedOutput: &app.Config{
 				Protocols: map[string]*app.Protocol{
@@ -63,32 +80,15 @@ func TestGetConfig(t *testing.T) {
 				},
 			},
 		},
-		"Missing CONFIG": {
-			Config:         "",
-			ExpectedError:  "CONFIG environment variable is required",
-			ExpectedOutput: &app.Config{},
-		},
-		"Invalid JSON": {
-			Config:         `{"protocols":}`,
-			ExpectedError:  "error parsing CONFIG: invalid character '}' looking for beginning of value",
-			ExpectedOutput: &app.Config{},
-		},
-		"Empty Protocols Section": {
-			Config: `{
-	"rules": [
-		{"cidr": "10.0.0.0/16"}
-	]
-}`,
-			ExpectedError:  "malformed configuration",
-			ExpectedOutput: &app.Config{},
-		},
-		"Empty Rules Section": {
-			Config: `{
-	"protocols": {
-		"http": {"transport": "tcp", "from_port": 80, "to_port": 80}
-	}
-}`,
-			ExpectedError:  "malformed configuration",
+		"Failure": {
+			Parameter: secret,
+			MockInput: &secretsmanager.GetSecretValueInput{
+				SecretId:     aws.String(secret),
+				VersionStage: aws.String("AWSCURRENT"),
+			},
+			MockOutput:     &secretsmanager.GetSecretValueOutput{},
+			MockError:      errors.New("reason"),
+			ExpectedError:  "error fetching secret: reason",
 			ExpectedOutput: &app.Config{},
 		},
 	}
@@ -98,9 +98,10 @@ func TestGetConfig(t *testing.T) {
 		counter++
 		t.Logf("Test Case %v/%v - %s", counter, len(suite), name)
 
-		t.Setenv("CONFIG", tc.Config)
+		m := new(mocks.SM)
+		m.On("GetSecretValue", tc.MockInput).Return(tc.MockOutput, tc.MockError).Once()
 
-		result, err := app.GetConfig()
+		result, err := app.GetConfig(m, tc.Parameter)
 
 		if tc.ExpectedError != "" {
 			assert.EqualError(err, tc.ExpectedError)